requirements.txt
aioredis
contextvars
discord.py
influxdb-client[async]
motor
pymongo
pyyaml
